+++ pagebuilder/src/Utils/Settings.js
@@ -15,3 +15,19 @@ export let findSection = (page, id, layer) => {
       return section;
     });
   }
+
+
+  // function to find section and add layer into it
+
+export let findSettingSection = (page, obj) => {
+    return page.map((section) => {
+      if (section.id === obj.id) {
+          //console.log('this is seetingggg -->', section)
+          section.containerType = obj.containerType
+          section.layers.push({...obj.layer, icon:""})
+      } else if (section.layers) {
+        section.layers = findSettingSection(section.layers);
+      }
+      return section;
+    });
+  }
